Set Windows Build

+++ src/Input.c
@@ -11,24 +11,24 @@ void mouseCallback(GLFWwindow* window, double xpos, double ypos)
 {
 	if (mouseFirstEnter)
 	{
-		g_MInput.lastX = xpos;
-		g_MInput.lastY = ypos;
+		g_MInput.lastX = (float)xpos;
+		g_MInput.lastY = (float)ypos;
 		g_MInput.yaw = -90.0f;
 		mouseFirstEnter = false;
 	}
 
 	double xOffset = xpos - g_MInput.lastX;
 	double yOffset = ypos - g_MInput.lastY;
-	g_MInput.lastX = xpos;
-	g_MInput.lastY = ypos;
+	g_MInput.lastX = (float)xpos;
+	g_MInput.lastY = (float)ypos;
 
-	g_MInput.yaw += xOffset;
-	g_MInput.pitch += yOffset;
+	g_MInput.yaw += (float)xOffset;
+	g_MInput.pitch += (float)yOffset;
 
 	vec3s direction;
-	direction.x = cos(glm_rad(g_MInput.yaw)) * cos(glm_rad(g_MInput.pitch));
-	direction.y = -sin(glm_rad(g_MInput.pitch));
-	direction.z = sin(glm_rad(g_MInput.yaw)) * cos(glm_rad(g_MInput.pitch));
+	direction.x = (float)(cos(glm_rad(g_MInput.yaw)) * cos(glm_rad(g_MInput.pitch)));
+	direction.y = (float)(-sin(glm_rad(g_MInput.pitch)));
+	direction.z = (float)(sin(glm_rad(g_MInput.yaw)) * cos(glm_rad(g_MInput.pitch)));
 	g_MainCamera.front = glms_vec3_normalize(direction);
 	
 	if(g_MInput.pitch > 89.0f) g_MInput.pitch =  89.0f;
